tidy(bots): switched from using ints to floats
RIP my beautiful integer mostly integer math :(.

+++ bot_src/c/minimal_player_info.h
@@ -1,17 +1,15 @@
 #ifndef MINIMAL_PLAYER_INFO_H
 #define MINIMAL_PLAYER_INFO_H
-#include <stdint.h>
-
 typedef struct MinimalPlayerInfo {
-	uint16_t pos_x;
-	uint16_t pos_y;
+	float pos_x;
+	float pos_y;
 	float direction;
 
 } MinimalPlayerInfo;
 
 typedef struct MinimalProjectileInfo {
-	uint16_t pos_x;
-	uint16_t pos_y;
+	float pos_x;
+	float pos_y;
 
 } MinimalProjectileInfo;
 #endif
